backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic>=2.12.0
python-dotenv==1.0.1
google-genai
github-copilot-sdk
openai>=1.30.0
boto3>=1.34.0
